add online extension editor and publishing
This lets users interactively create and publish extensions on the web app, editing the JavaScript source code in a Monaco editor. There is no TypeScript compilation or dependency packaging provided, so this is quite limited. But it supports some neat demos.

TODO!(sqs): make this feature-flagged or at least indicate the limitations inline